Correct sign mismatch in Visual C++ 2019, take 2
Also an erratum for printf of a size_t variable

+++ ARGLINK_REWRITE/Program.cs
@@ -97,7 +97,7 @@ Note: DOS has a 126-char limit on parameters, so please use the @ option.
 		private static int Search(List<LinkData> link, string name)
 		{
 			int nameId = -1;
-			for (int i = 0; i < link.Count; i++) {
+			for (int i = 0; i < (int)link.Count; i++) {	// Cast for MSVC
 				if (link[i].Name.Equals(name)) {
 					nameId = i;
 				}
@@ -351,7 +351,7 @@ Note: DOS has a 126-char limit on parameters, so please use the @ option.
 							int highestdeep    = -1;
 							int highestdeepidx = -1;
 							int i;
-							for (i = 1; i < linkcalc.Count; i++) {
+							for (i = 1; i < (int)linkcalc.Count; i++) {	// Cast for MSVC
 								//Get the first highest one
 								if (highestdeep < linkcalc[i].Deep) {
 									highestdeep    = linkcalc[i].Deep;
@@ -362,7 +362,7 @@ Note: DOS has a 126-char limit on parameters, so please use the @ option.
 							//Check for highest priority
 							int highestpri    = -1;
 							int highestpriidx = -1;
-							for (i = highestdeepidx; i < linkcalc.Count; i++) {
+							for (i = highestdeepidx; i < (int)linkcalc.Count; i++) {	// Cast for MSVC
 								//Get the first highest one
 								if (linkcalc[i].Deep != highestdeep || highestpri > linkcalc[i].Priority) {
 									break;
